+++ lib/libutils/ext/ftrace/ftrace.c
@@ -32,7 +32,7 @@ static const char hex_str[] = "0123456789abcdef";
 static __noprof struct ftrace_buf *get_fbuf(void)
 {
 #if defined(__KERNEL__)
-	int ct = thread_get_id_may_fail();
+	short int ct = thread_get_id_may_fail();
 	struct tee_ta_session *s = NULL;
 	struct thread_specific_data *tsd = NULL;
